requirements.txt
pygame
pyyaml
numpy
matplotlib
sympy
scipy
torch